+++ lib/Features/Auth/Login/presentation/view/ForgetPasswordScreen.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:online_exam/Core/widgets/custom_app_bar.dart';
-import 'package:online_exam/Features/Auth/Login/presentation/view/widgets/forget_password_view_body.dart';
+import 'package:online_exam/Features/Auth/Login/presentation/view/widgets/ForgetPasswordScreenBody.dart';
 
-class ForgetPasswordView extends StatelessWidget {
-  const ForgetPasswordView({super.key});
+class ForgetPasswordScreen extends StatelessWidget {
+  const ForgetPasswordScreen({super.key});
   static String id = 'ForgetPassword';
   @override
   Widget build(BuildContext context) {
@@ -16,7 +16,7 @@ class ForgetPasswordView extends StatelessWidget {
           leadingVisibility: true,
         ),
       ),
-      body: ForgetPasswordViewBody(),
+      body: ForgetPasswordScreenBody(),
     );
   }
 }
