+++ lib/main.dart
@@ -4,7 +4,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:online_exam/Core/di/di.dart';
 import 'package:online_exam/Core/utils/app_routes.dart';
-import 'package:online_exam/Features/Auth/Login/presentation/view/login_screen.dart';
+import 'package:online_exam/Features/Auth/Login/presentation/view/LoginScreen.dart';
 
 void main() {
   configureDependencies();
