+++ pubspec.yaml
@@ -41,6 +41,7 @@ dependencies:
   json_annotation: ^4.9.0
   flutter_screenutil: ^5.9.3
   google_fonts: ^6.2.1
+  dartz: ^0.10.1
 
 dev_dependencies:
   flutter_test:
